Support optional when generate grpc stub js code
