+++ package.json
@@ -60,7 +60,7 @@
     "@chatie/tsconfig": "^0.14.1",
     "@types/google-protobuf": "^3.7.4",
     "@types/protobufjs": "^6.0.0",
-    "grpc-tools": "^1.11.0",
+    "grpc-tools": "^1.11.1",
     "grpc_tools_node_protoc_ts": "^5.1.3",
     "grpcc": "^1.1.3",
     "nodemon": "^2.0.7",
